+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Alter table: Do not highlight other move handles while dragging a column
 - Alter table: Move columns on touch screens
 - Alter table: Edit multiline column comments in a textarea
+- Alter table, Privileges: Do not print a success message if nothing was changed
 - Select: Fix the number of selected rows after going back in history
 - Edit: Keep values of columns with special characters in name after error
 - Edit: Fix switching function to SQL with the edit-foreign plugin (bug #1254)